+++ packages/terra-clinical-detail-view/src/terra-dev-site/doc/clinical-detail-view/clinicalDetailView.1.doc.jsx
@@ -7,6 +7,9 @@ import { name } from '../../../../package.json';
 import DetailViewSrc from '!raw-loader!../../../../src/DetailView';
 import DetailListSrc from '!raw-loader!../../../../src/DetailList';
 import DetailListItemSrc from '!raw-loader!../../../../src/DetailListItem';
+import DetailViewPropsTable from '!terra-props-table-loader!../../../../src/DetailView';
+import DetailListPropsTable from '!terra-props-table-loader!../../../../src/DetailList';
+import DetailListItemPropsTable from '!terra-props-table-loader!../../../../src/DetailListItem';
 
 // Example files
 import DetailViewDivided from '../example/DetailViewDivided';
@@ -42,14 +45,17 @@ const DocPage = () => (
       {
         componentName: 'Detail View',
         componentSrc: DetailViewSrc,
+        componentProps: DetailViewPropsTable,
       },
       {
         componentName: 'Detail List',
         componentSrc: DetailListSrc,
+        componentProps: DetailListPropsTable,
       },
       {
         componentName: 'Detail List Item',
         componentSrc: DetailListItemSrc,
+        componentProps: DetailListItemPropsTable,
       },
     ]}
   />